rename files to follow facebook's conventions
