+++ packages/jest-cli/src/ReporterDispatcher.js
