+++ packages/jest-cli/src/SnapshotInteractiveMode.js
