+++ packages/jest-cli/src/TestWatcher.js
