+++ packages/jest-cli/src/__tests__/watch_test_name_pattern_mode.test.js
@@ -24,7 +24,7 @@ jest.mock('ansi-escapes', () => ({
 }));
 
 jest.mock(
-  '../search_source',
+  '../SearchSource',
   () =>
     class {
       findMatchingTests(pattern) {
@@ -39,7 +39,7 @@ jest.doMock('strip-ansi');
 require('strip-ansi').mockImplementation(str => str);
 
 jest.doMock(
-  '../run_jest',
+  '../runJest',
   () =>
     function() {
       const args = Array.from(arguments);
