+++ packages/jest-cli/src/collectHandles.js
