+++ packages/jest-cli/src/generateEmptyCoverage.js
