+++ packages/jest-cli/src/preRunMessage.js
