+++ packages/jest-cli/src/reporters/base_reporter.js
@@ -12,7 +12,7 @@ import type {Context} from 'types/Context';
 import type {Test} from 'types/TestRunner';
 import type {ReporterOnStartOptions} from 'types/Reporters';
 
-import {remove as preRunMessageRemove} from '../pre_run_message';
+import {remove as preRunMessageRemove} from '../preRunMessage';
 
 export default class BaseReporter {
   _error: ?Error;
