+++ packages/jest-cli/src/testPathPatternToRegexp.js
