+++ packages/jest-cli/src/testResultHelpers.js
