+++ packages/jest-cli/src/watch.js
@@ -14,18 +14,18 @@ import type {Options as UpdateGlobalConfigOptions} from './lib/update_global_con
 
 import ansiEscapes from 'ansi-escapes';
 import chalk from 'chalk';
-import getChangedFilesPromise from './get_changed_files_promise';
+import getChangedFilesPromise from './getChangedFilesPromise';
 import exit from 'exit';
 import HasteMap from 'jest-haste-map';
 import isValidPath from './lib/is_valid_path';
 import {isInteractive} from 'jest-util';
-import {print as preRunMessagePrint} from './pre_run_message';
+import {print as preRunMessagePrint} from './preRunMessage';
 import createContext from './lib/create_context';
-import runJest from './run_jest';
+import runJest from './runJest';
 import updateGlobalConfig from './lib/update_global_config';
-import SearchSource from './search_source';
-import TestWatcher from './test_watcher';
-import FailedTestsCache from './failed_tests_cache';
+import SearchSource from './SearchSource';
+import TestWatcher from './TestWatcher';
+import FailedTestsCache from './FailedTestsCache';
 import {CLEAR} from './constants';
 import {KEYS, JestHook} from 'jest-watcher';
 import TestPathPatternPlugin from './plugins/test_path_pattern';
